Clean internal structure of `typhon.utils`.
